+++ projects/FieldNation/fieldnation/src/main/java/com/fieldnation/service/data/documents/DocumentDispatch.java
@@ -13,10 +13,10 @@ import java.io.File;
  */
 public class DocumentDispatch implements DocumentConstants {
 
-    public static void download(Context context, long documentId, File file, int state, boolean isSync) {
+    public static void download(Context context, int documentId, File file, int state, boolean isSync) {
         Bundle bundle = new Bundle();
         bundle.putString(PARAM_ACTION, PARAM_ACTION_DOWNLOAD_DOCUMENT);
-        bundle.putLong(PARAM_DOCUMENT_ID, documentId);
+        bundle.putInt(PARAM_DOCUMENT_ID, documentId);
         bundle.putBoolean(PARAM_IS_SYNC, isSync);
 
         bundle.putInt(PARAM_STATE, state);
